Move arrow interator into script
